+++ client/web/src/enterprise/campaigns/close/CampaignClosePage.story.tsx
@@ -32,26 +32,19 @@ const now = new Date()
 
 const campaignDefaults: CampaignFields = {
     __typename: 'Campaign',
-    changesets: {
-        stats: {
-            closed: 1,
-            deleted: 1,
-            merged: 2,
-            draft: 1,
-            open: 2,
-            total: 10,
-            unpublished: 4,
-        },
+    changesetsStats: {
+        closed: 1,
+        deleted: 1,
+        merged: 2,
+        draft: 1,
+        open: 2,
+        total: 10,
+        unpublished: 4,
     },
     createdAt: subDays(now, 5).toISOString(),
     initialApplier: {
         url: '/users/alice',
         username: 'alice',
-    },
-    diffStat: {
-        added: 10,
-        changed: 8,
-        deleted: 10,
     },
     id: 'specid',
     url: '/users/alice/campaigns/specid',