+++ enterprise/internal/codeintel/api/locations.go
@@ -1,7 +1,7 @@
 package api
 
 import (
-	bundles "github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/bundles/client"
+	bundles "github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/bundles/client_types"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/store"
 )