+++ internal/extsvc/github/repos.go
@@ -39,41 +39,6 @@ type Repository struct {
 	ViewerPermission string // ADMIN, WRITE, READ, or empty if unknown. Only the graphql api populates this. https://developer.github.com/v4/enum/repositorypermission/
 }
 
-// repositoryFieldsGraphQLFragment returns a GraphQL fragment that contains the fields needed to populate the
-// Repository struct.
-func (c *Client) repositoryFieldsGraphQLFragment() string {
-	if c.githubDotCom {
-		return `
-fragment RepositoryFields on Repository {
-	id
-	databaseId
-	nameWithOwner
-	description
-	url
-	isPrivate
-	isFork
-	isArchived
-	viewerPermission
-}
-	`
-	}
-	// Some fields are not yet available on GitHub Enterprise yet
-	// or are available but too new to expect our customers to have updated:
-	// - viewerPermission
-	return `
-fragment RepositoryFields on Repository {
-	id
-	databaseId
-	nameWithOwner
-	description
-	url
-	isPrivate
-	isFork
-	isArchived
-}
-	`
-}
-
 func ownerNameCacheKey(owner, name string) string       { return "0:" + owner + "/" + name }
 func nameWithOwnerCacheKey(nameWithOwner string) string { return "0:" + nameWithOwner }
 func nodeIDCacheKey(id string) string                   { return "1:" + id }
@@ -89,7 +54,7 @@ func MockGetRepository_Return(returns *Repository) {
 }
 
 // GetRepository gets a repository from GitHub by owner and repository name.
-func (c *Client) GetRepository(ctx context.Context, owner, name string) (*Repository, error) {
+func (c *V3Client) GetRepository(ctx context.Context, owner, name string) (*Repository, error) {
 	if GetRepositoryMock != nil {
 		return GetRepositoryMock(ctx, owner, name)
 	}
@@ -105,33 +70,8 @@ func (c *Client) GetRepository(ctx context.Context, owner, name string) (*Reposi
 	}, false)
 }
 
-// GetRepositoryByNodeID gets a repository from GitHub by its GraphQL node ID using the specified user token.
-func (c *Client) GetRepositoryByNodeID(ctx context.Context, id string) (*Repository, error) {
-	return c.getRepositoryByNodeID(ctx, id, false)
-}
-
-// HACK: allows us to bypass the GitHub client cache (used for fetching repositories to determine
-// permissions). TODO(beyang): merge this into a unified GetRepositoryByNodeID with a clean
-// interface.
-func (c *Client) GetRepositoryByNodeIDNoCache(ctx context.Context, id string) (*Repository, error) {
-	return c.getRepositoryByNodeID(ctx, id, true)
-}
-
-func (c *Client) getRepositoryByNodeID(ctx context.Context, id string, nocache bool) (*Repository, error) {
-	key := nodeIDCacheKey(id)
-
-	return c.cachedGetRepository(ctx, key, func(ctx context.Context) (repo *Repository, keys []string, err error) {
-		keys = append(keys, key)
-		repo, err = c.getRepositoryByNodeIDFromAPI(ctx, id)
-		if repo != nil {
-			keys = append(keys, nameWithOwnerCacheKey(repo.NameWithOwner)) // also cache under "owner/name"
-		}
-		return repo, keys, err
-	}, nocache)
-}
-
 // cachedGetRepository caches the getRepositoryFromAPI call.
-func (c *Client) cachedGetRepository(ctx context.Context, key string, getRepositoryFromAPI func(ctx context.Context) (repo *Repository, keys []string, err error), nocache bool) (*Repository, error) {
+func (c *V3Client) cachedGetRepository(ctx context.Context, key string, getRepositoryFromAPI func(ctx context.Context) (repo *Repository, keys []string, err error), nocache bool) (*Repository, error) {
 	if !nocache {
 		if cached := c.getRepositoryFromCache(ctx, key); cached != nil {
 			reposGitHubCacheCounter.WithLabelValues("hit").Inc()
@@ -178,7 +118,7 @@ type cachedRepo struct {
 
 // getRepositoryFromCache attempts to get a response from the redis cache.
 // It returns nil error for cache-hit condition and non-nil error for cache-miss.
-func (c *Client) getRepositoryFromCache(ctx context.Context, key string) *cachedRepo {
+func (c *V3Client) getRepositoryFromCache(ctx context.Context, key string) *cachedRepo {
 	b, ok := c.repoCache.Get(strings.ToLower(key))
 	if !ok {
 		return nil
@@ -195,7 +135,7 @@ func (c *Client) getRepositoryFromCache(ctx context.Context, key string) *cached
 // addRepositoryToCache will cache the value for repo. The caller can provide multiple cache keys
 // for the multiple ways that this repository can be retrieved (e.g., both "owner/name" and the
 // GraphQL node ID).
-func (c *Client) addRepositoryToCache(keys []string, repo *cachedRepo) {
+func (c *V3Client) addRepositoryToCache(keys []string, repo *cachedRepo) {
 	b, err := json.Marshal(repo)
 	if err != nil {
 		return
@@ -207,7 +147,7 @@ func (c *Client) addRepositoryToCache(keys []string, repo *cachedRepo) {
 
 // addRepositoriesToCache will cache repositories that exist
 // under relevant cache keys.
-func (c *Client) addRepositoriesToCache(repos []*Repository) {
+func (c *V3Client) addRepositoriesToCache(repos []*Repository) {
 	for _, repo := range repos {
 		keys := []string{nameWithOwnerCacheKey(repo.NameWithOwner), nodeIDCacheKey(repo.ID)} // cache under multiple
 		c.addRepositoryToCache(keys, &cachedRepo{Repository: *repo})
@@ -233,7 +173,7 @@ type restRepository struct {
 }
 
 // getRepositoryFromAPI attempts to fetch a repository from the GitHub API without use of the redis cache.
-func (c *Client) getRepositoryFromAPI(ctx context.Context, owner, name string) (*Repository, error) {
+func (c *V3Client) getRepositoryFromAPI(ctx context.Context, owner, name string) (*Repository, error) {
 	// If no token, we must use the older REST API, not the GraphQL API. See
 	// https://platform.github.community/t/anonymous-access/2093/2. This situation occurs on (for
 	// example) a server with autoAddRepos and no GitHub connection configured when someone visits
@@ -284,88 +224,12 @@ func convertRestRepoPermissions(restRepoPermissions restRepositoryPermissions) s
 // An empty sinceRepoID returns the first page of results.
 // This is only intended to be called for GitHub Enterprise, so no rate limit information is returned.
 // https://developer.github.com/v3/repos/#list-all-public-repositories
-func (c *Client) getPublicRepositories(ctx context.Context, sinceRepoID int64) ([]*Repository, error) {
+func (c *V3Client) getPublicRepositories(ctx context.Context, sinceRepoID int64) ([]*Repository, error) {
 	path := "repositories"
 	if sinceRepoID > 0 {
 		path += "?per_page=100&since=" + strconv.FormatInt(sinceRepoID, 10)
 	}
 	return c.listRepositories(ctx, path)
-}
-
-// getRepositoryByNodeIDFromAPI attempts to fetch a repository by GraphQL node ID from the GitHub
-// API without use of the redis cache.
-func (c *Client) getRepositoryByNodeIDFromAPI(ctx context.Context, id string) (*Repository, error) {
-	var result struct {
-		Node *Repository `json:"node"`
-	}
-	if err := c.requestGraphQL(ctx, `
-query Repository($id: ID!) {
-	node(id: $id) {
-		... on Repository {
-			...RepositoryFields
-		}
-	}
-}`+c.repositoryFieldsGraphQLFragment(),
-		map[string]interface{}{"id": id},
-		&result,
-	); err != nil {
-		if gqlErrs, ok := err.(graphqlErrors); ok {
-			for _, err2 := range gqlErrs {
-				if err2.Type == graphqlErrTypeNotFound {
-					return nil, ErrNotFound
-				}
-			}
-		}
-		return nil, err
-	}
-	if result.Node == nil {
-		return nil, ErrNotFound
-	}
-	return result.Node, nil
-}
-
-// MaxNodeIDs is the maximum number of repository nodes that can be queried in one call to the
-// GitHub GraphQL API.
-var MaxNodeIDs = 100
-
-// GetRepositoriesByNodeIDFromAPI fetches the specified repositories (nodeIDs) and returns a map
-// from node ID to repository metadata. If a repository is not found, it will not be present in the
-// return map. The caller should respect the max nodeID count limit of the GitHub API, which at the
-// time of writing, is 100 (if the caller does not respect this match, this method will return an
-// error). This method does not cache.
-func (c *Client) GetRepositoriesByNodeIDFromAPI(ctx context.Context, nodeIDs []string) (map[string]*Repository, error) {
-	var result struct {
-		Nodes []*Repository
-	}
-	err := c.requestGraphQL(ctx, `
-query Repositories($ids: [ID!]!) {
-	nodes(ids: $ids) {
-		... on Repository {
-			...RepositoryFields
-		}
-	}
-}
-`+c.repositoryFieldsGraphQLFragment(), map[string]interface{}{"ids": nodeIDs}, &result)
-	if err != nil {
-		if gqlErrs, ok := err.(graphqlErrors); ok {
-			for _, err2 := range gqlErrs {
-				if err2.Type == graphqlErrTypeNotFound {
-					continue
-				}
-				return nil, err
-			}
-		} else {
-			return nil, err
-		}
-	}
-
-	repos := make(map[string]*Repository)
-	for _, r := range result.Nodes {
-		if r != nil {
-			repos[r.ID] = r
-		}
-	}
-	return repos, nil
 }
 
 // ErrBatchTooLarge is when the requested batch of GitHub repositories to fetch
@@ -384,7 +248,7 @@ var ErrBatchTooLarge = errors.New("requested batch of GitHub repositories too la
 // the conservative step back from 37.
 //
 // This method does not cache.
-func (c *Client) GetReposByNameWithOwner(ctx context.Context, namesWithOwners ...string) ([]*Repository, error) {
+func (c *V4Client) GetReposByNameWithOwner(ctx context.Context, namesWithOwners ...string) ([]*Repository, error) {
 	if len(namesWithOwners) > 30 {
 		return nil, ErrBatchTooLarge
 	}
@@ -419,7 +283,7 @@ func (c *Client) GetReposByNameWithOwner(ctx context.Context, namesWithOwners ..
 	return repos, nil
 }
 
-func (c *Client) buildGetReposBatchQuery(namesWithOwners []string) (string, error) {
+func (c *V4Client) buildGetReposBatchQuery(namesWithOwners []string) (string, error) {
 	var b strings.Builder
 	b.WriteString(c.repositoryFieldsGraphQLFragment())
 	b.WriteString("query {\n")
@@ -438,7 +302,42 @@ func (c *Client) buildGetReposBatchQuery(namesWithOwners []string) (string, erro
 	return b.String(), nil
 }
 
-func (c *Client) ListPublicRepositories(ctx context.Context, sinceRepoID int64) ([]*Repository, error) {
+// repositoryFieldsGraphQLFragment returns a GraphQL fragment that contains the fields needed to populate the
+// Repository struct.
+func (c *V4Client) repositoryFieldsGraphQLFragment() string {
+	if c.githubDotCom {
+		return `
+fragment RepositoryFields on Repository {
+	id
+	databaseId
+	nameWithOwner
+	description
+	url
+	isPrivate
+	isFork
+	isArchived
+	viewerPermission
+}
+	`
+	}
+	// Some fields are not yet available on GitHub Enterprise yet
+	// or are available but too new to expect our customers to have updated:
+	// - viewerPermission
+	return `
+fragment RepositoryFields on Repository {
+	id
+	databaseId
+	nameWithOwner
+	description
+	url
+	isPrivate
+	isFork
+	isArchived
+}
+	`
+}
+
+func (c *V3Client) ListPublicRepositories(ctx context.Context, sinceRepoID int64) ([]*Repository, error) {
 	repos, err := c.getPublicRepositories(ctx, sinceRepoID)
 	if err != nil {
 		return nil, err
@@ -459,7 +358,7 @@ const (
 // ListAffiliatedRepositories lists GitHub repositories affiliated with the client
 // token. page is the page of results to return. Pages are 1-indexed (so the
 // first call should be for page 1).
-func (c *Client) ListAffiliatedRepositories(ctx context.Context, visibility Visibility, page int) (
+func (c *V3Client) ListAffiliatedRepositories(ctx context.Context, visibility Visibility, page int) (
 	repos []*Repository,
 	hasNextPage bool,
 	rateLimitCost int,
@@ -477,7 +376,7 @@ func (c *Client) ListAffiliatedRepositories(ctx context.Context, visibility Visi
 // ListOrgRepositories lists GitHub repositories from the specified organization.
 // org is the name of the organization. page is the page of results to return.
 // Pages are 1-indexed (so the first call should be for page 1).
-func (c *Client) ListOrgRepositories(ctx context.Context, org string, page int) (repos []*Repository, hasNextPage bool, rateLimitCost int, err error) {
+func (c *V3Client) ListOrgRepositories(ctx context.Context, org string, page int) (repos []*Repository, hasNextPage bool, rateLimitCost int, err error) {
 	path := fmt.Sprintf("orgs/%s/repos?sort=created&page=%d&per_page=100", org, page)
 	repos, err = c.listRepositories(ctx, path)
 	return repos, len(repos) > 0, 1, err
@@ -485,7 +384,7 @@ func (c *Client) ListOrgRepositories(ctx context.Context, org string, page int) 
 
 // ListUserRepositories lists GitHub repositories from the specified user.
 // Pages are 1-indexed (so the first call should be for page 1)
-func (c *Client) ListUserRepositories(ctx context.Context, user string, page int) (repos []*Repository, hasNextPage bool, rateLimitCost int, err error) {
+func (c *V3Client) ListUserRepositories(ctx context.Context, user string, page int) (repos []*Repository, hasNextPage bool, rateLimitCost int, err error) {
 	path := fmt.Sprintf("users/%s/repos?sort=created&type=owner&page=%d&per_page=100", user, page)
 	repos, err = c.listRepositories(ctx, path)
 	return repos, len(repos) > 0, 1, err
@@ -504,7 +403,7 @@ type RepositoryListPage struct {
 	HasNextPage bool
 }
 
-func (c *Client) ListRepositoriesForSearch(ctx context.Context, searchString string, page int) (RepositoryListPage, error) {
+func (c *V3Client) ListRepositoriesForSearch(ctx context.Context, searchString string, page int) (RepositoryListPage, error) {
 	urlValues := url.Values{
 		"q":        []string{searchString},
 		"page":     []string{strconv.Itoa(page)},
@@ -536,7 +435,7 @@ type restTopicsResponse struct {
 }
 
 // ListTopicsOnRepository lists topics on the given repository.
-func (c *Client) ListTopicsOnRepository(ctx context.Context, ownerAndName string) ([]string, error) {
+func (c *V3Client) ListTopicsOnRepository(ctx context.Context, ownerAndName string) ([]string, error) {
 	owner, name, err := SplitRepositoryNameWithOwner(ownerAndName)
 	if err != nil {
 		return nil, err
@@ -550,4 +449,41 @@ func (c *Client) ListTopicsOnRepository(ctx context.Context, ownerAndName string
 		return nil, err
 	}
 	return result.Names, nil
+}
+
+// ListInstallationRepositories lists repositories on which the authenticated
+// GitHub App has been installed.
+func (c *V3Client) ListInstallationRepositories(ctx context.Context) ([]*Repository, error) {
+	type response struct {
+		Repositories []restRepository `json:"repositories"`
+	}
+	var resp response
+	if err := c.requestGet(ctx, "installation/repositories", &resp); err != nil {
+		return nil, err
+	}
+	repos := make([]*Repository, 0, len(resp.Repositories))
+	for _, restRepo := range resp.Repositories {
+		repos = append(repos, convertRestRepo(restRepo))
+	}
+	return repos, nil
+}
+
+// listRepositories is a generic method that unmarshals the given
+// JSON HTTP endpoint into a []restRepository. It will return an
+// error if it fails.
+//
+// This is used to extract repositories from the GitHub API endpoints:
+// - /users/:user/repos
+// - /orgs/:org/repos
+// - /user/repos
+func (c *V3Client) listRepositories(ctx context.Context, requestURI string) ([]*Repository, error) {
+	var restRepos []restRepository
+	if err := c.requestGet(ctx, requestURI, &restRepos); err != nil {
+		return nil, err
+	}
+	repos := make([]*Repository, 0, len(restRepos))
+	for _, restRepo := range restRepos {
+		repos = append(repos, convertRestRepo(restRepo))
+	}
+	return repos, nil
 }